use int instead of int64 for PartNumber.

+++ s3/pkg/model/xmlstruct.go
@@ -45,14 +45,16 @@ type InitiateMultipartUploadResult struct {
 	UploadId string `xml:"UploadId"`
 }
 
+//PartNumber should be between 1 and 10000.
+//Please refer to https://docs.aws.amazon.com/zh_cn/AmazonS3/latest/dev/qfacts.html
 type UploadPartResult struct {
 	Xmlns      string `xml:"xmlns,attr"`
-	PartNumber int64  `xml:"PartNumber"`
+	PartNumber int    `xml:"PartNumber"`
 	ETag       string `xml:"ETag"`
 }
 
 type Part struct {
-	PartNumber int64  `xml:"PartNumber"`
+	PartNumber int    `xml:"PartNumber"`
 	ETag       string `xml:"ETag"`
 }
